requirements.txt
certifi
charset-normalizer
coloredlogs
colorzero
contourpy
cycler
filelock
flatbuffers
fonttools
fsspec
gpiozero
humanfriendly
idna
Jinja2
kiwisolver
lgpio
MarkupSafe
matplotlib
mpmath
networkx
numpy
onnx
onnxruntime
opencv-python
packaging
pandas
pillow
pip
protobuf
psutil
py-cpuinfo
pyparsing
python-dateutil
pytz
PyYAML
requests
RPi.GPIO
scipy
seaborn
setuptools
six
sympy
torch
torchvision
tqdm
typing_extensions
tzdata
ultralytics-thop
urllib3
